Revert "Refactor and implement OrganizationTree and PersonNode components"
