--- OrganizationTreeV2/components/core/index.ts
@@ -1,3 +1,0 @@
-export { OrganizationTree } from "./OrganizationTree";
-export type { OrganizationTreeProps } from "./OrganizationTree";
-export { useOrganizationTreeStyles } from "./OrganizationTree.styles";
