--- OrganizationTreeV2/components/index.ts
@@ -1,4 +1,0 @@
-export { OrganizationTree } from "./core/OrganizationTree";
-export { PersonNode } from "./nodes/PersonNode";
-export type { OrganizationTreeProps } from "./core/OrganizationTree";
-export type { PersonNodeProps } from "./nodes/PersonNode";
