--- OrganizationTreeV2/components/nodes/index.ts
@@ -1,3 +1,0 @@
-export { PersonNode } from "./PersonNode";
-export type { PersonNodeProps } from "./PersonNode";
-export * from "./PersonNode.logic";
