--- OrganizationTreeV2/components/panels/index.ts
@@ -1,6 +1,0 @@
-export { ReactFlowContent } from "./ReactFlowContent";
-export { SurveyPanel } from "./SurveyPanel";
-export { DescriptionPanel } from "./DescriptionPanel";
-export type { ReactFlowContentProps } from "./ReactFlowContent";
-export type { SurveyPanelProps } from "./SurveyPanel";
-export type { DescriptionPanelProps } from "./DescriptionPanel";
