--- OrganizationTreeV2/services/data/index.ts
@@ -1,2 +1,0 @@
-export { OrganizationService } from "./OrganizationService";
-export * from "./OrganizationService.internal";
