--- OrganizationTreeV2/services/index.ts
@@ -1,3 +1,0 @@
-export { OrganizationService } from "./data/OrganizationService";
-export { LayoutService } from "./layout/LayoutService";
-export * from "./utils/constants";
