--- OrganizationTreeV2/services/layout/index.ts
@@ -1,1 +1,0 @@
-export { LayoutService } from "./LayoutService";
