--- OrganizationTreeV2/services/utils/constants.ts
@@ -1,17 +1,0 @@
-/**
- * Constants used throughout the OrganizationTreeV2 control
- */
-
-// Layout constants
-export const FIXED_HEIGHT = 768;
-
-// Dagre layout constants
-export const LAYOUT_CONSTANTS = {
-  NODE_WIDTH: 220,
-  NODE_HEIGHT: 140,
-  RANK_SEPARATION: 80,
-  NODE_SEPARATION: 60,
-  EDGE_SEPARATION: 10,
-  MARGIN_X: 20,
-  MARGIN_Y: 20,
-} as const;
