--- OrganizationTreeV2/types/index.ts
@@ -1,9 +1,0 @@
-export type {
-  OrganizationPerson,
-  OrganizationNode,
-  OrganizationEdge,
-  HierarchyFilterOptions,
-  SurveyResponse,
-  Survey,
-  SelectedSurvey,
-} from "./OrganizationTypes";
