+++ package.json
@@ -10,9 +10,7 @@
     "rebuild": "pcf-scripts rebuild",
     "start": "pcf-scripts start",
     "start:watch": "pcf-scripts start watch",
-    "refreshTypes": "pcf-scripts refreshTypes",
-    "format": "prettier --write \"OrganizationTreeV2/**/*.{ts,tsx,js,jsx,json}\" \"*.{ts,tsx,js,jsx,json,mjs}\"",
-    "format:check": "prettier --check \"OrganizationTreeV2/**/*.{ts,tsx,js,jsx,json}\" \"*.{ts,tsx,js,jsx,json,mjs}\""
+    "refreshTypes": "pcf-scripts refreshTypes"
   },
   "dependencies": {
     "@codaworks/react-glow": "^1.0.6",
@@ -34,7 +32,6 @@
     "globals": "^15.15.0",
     "pcf-scripts": "^1",
     "pcf-start": "^1",
-    "prettier": "^3.6.2",
     "react": "^17.0.2",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.31.0"
