+++ pcfconfig.json
@@ -1,3 +1,3 @@
 {
-  "outDir": "./out/controls"
+    "outDir": "./out/controls"
 }
