+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
-  "extends": "./node_modules/pcf-scripts/tsconfig_base.json",
-  "compilerOptions": {
-    "typeRoots": ["node_modules/@types"],
-    "allowSyntheticDefaultImports": true
-  }
+    "extends": "./node_modules/pcf-scripts/tsconfig_base.json",
+    "compilerOptions": {
+        "typeRoots": ["node_modules/@types"]
+    }
 }
